Extract database access and types
