+++ backend/sendmessage/src/login-user.ts
@@ -1,86 +1,23 @@
-import { CardValue, VOTE_NOTE_VOTED } from '../../../shared/cards';
-import { SCALES } from '../../../shared/scales';
-import { Config, ConfigWithHandler } from '../../shared/config';
-import { getGroupItem } from '../../shared/getGroupItem';
+import { VOTE_NOTE_VOTED } from '../../../shared/cards';
+import { addConnectionToGroup } from '../../shared/database/addConnectionToGroup';
+import { addUserAndGroupToConnection } from '../../shared/database/addUserAndGroupToConnection';
+import { createGroupWithConnection } from '../../shared/database/createGroupWithConnection';
+import { getGroup } from '../../shared/database/getGroup';
 import { getTtl } from '../../shared/getTtl';
+import { ConfigWithHandler } from '../../shared/types';
 import { broadcastState } from './broadcast-state';
-
-const EXPIRY_TIME_IN_HOUR = process.env.EXPIRY_TIME_IN_HOUR || '16';
+import { EXPIRY_TIME_IN_HOUR } from './const';
 
 export const loginUser = async (userId: string, groupId: string, config: ConfigWithHandler) => {
-  const groupItem = await getGroupItem(groupId, config);
+  const groupItem = await getGroup(groupId, config);
   const groupUpdate = groupItem
-    ? addUserToGroup(
+    ? addConnectionToGroup(
         groupId,
         userId,
         groupItem.connections[userId]?.vote || VOTE_NOTE_VOTED,
         config
       )
-    : createGroupWithUser(groupId, userId, config);
+    : createGroupWithConnection(groupId, userId, getTtl(EXPIRY_TIME_IN_HOUR), config);
   await Promise.all([addUserAndGroupToConnection(groupId, userId, config), groupUpdate]);
   await broadcastState(groupId, config);
 };
-
-const addUserToGroup = (
-  groupId: string,
-  userId: string,
-  vote: CardValue,
-  { ddb, tableName, connectionId }: Config
-) =>
-  ddb
-    .update({
-      TableName: tableName,
-      Key: {
-        primaryKey: `groupId:${groupId}`,
-      },
-      UpdateExpression: `SET connections.#userId = :connection`,
-      ExpressionAttributeNames: {
-        '#userId': userId,
-      },
-      ExpressionAttributeValues: {
-        ':connection': {
-          connectionId,
-          vote,
-        },
-      },
-    })
-    .promise();
-
-const createGroupWithUser = (
-  groupId: string,
-  userId: string,
-  { ddb, tableName, connectionId }: Config
-) =>
-  ddb
-    .put({
-      TableName: tableName,
-      Item: {
-        primaryKey: `groupId:${groupId}`,
-        ttl: getTtl(EXPIRY_TIME_IN_HOUR),
-        connections: {
-          [userId]: { connectionId: connectionId, vote: VOTE_NOTE_VOTED },
-        },
-        groupId,
-        scale: SCALES.COHEN_SCALE.values,
-      },
-    })
-    .promise();
-
-const addUserAndGroupToConnection = (
-  groupId: string,
-  userId: string,
-  { ddb, tableName, connectionId }: Config
-) =>
-  ddb
-    .update({
-      TableName: tableName,
-      Key: {
-        primaryKey: `connectionId:${connectionId}`,
-      },
-      UpdateExpression: 'set userId = :userId, groupId = :groupId',
-      ExpressionAttributeValues: {
-        ':userId': userId,
-        ':groupId': groupId,
-      },
-    })
-    .promise();
