+++ backend/sendmessage/src/set-vote.ts
@@ -1,31 +1,15 @@
 import { CardValue } from '../../../shared/cards';
-import { Config, ConfigWithHandler } from '../../shared/config';
-import { getConnectionItem } from '../../shared/getConnectionItem';
+import { getConnection } from '../../shared/database/getConnection';
+import { setGroupConnectionVote } from '../../shared/database/setGroupConnectionVote';
+import { ConfigWithHandler } from '../../shared/types';
 import { broadcastState } from './broadcast-state';
 
 export const setVote = async (vote: CardValue, config: ConfigWithHandler) => {
-  const connectionItem = await getConnectionItem(config);
+  const connectionItem = await getConnection(config);
   if (!connectionItem) return;
   const { groupId, userId } = connectionItem;
   if (!(groupId && userId)) return;
 
-  await updateVote(groupId, userId, vote, config);
+  await setGroupConnectionVote(groupId, userId, vote, config);
   await broadcastState(groupId, config);
 };
-
-const updateVote = (groupId: string, userId: string, vote: CardValue, { ddb, tableName }: Config) =>
-  ddb
-    .update({
-      TableName: tableName,
-      Key: {
-        primaryKey: `groupId:${groupId}`,
-      },
-      UpdateExpression: `SET connections.#userId.vote = :vote`,
-      ExpressionAttributeNames: {
-        '#userId': userId,
-      },
-      ExpressionAttributeValues: {
-        ':vote': vote,
-      },
-    })
-    .promise();
