+++ backend/shared/database/deleteConnection.ts
@@ -1,0 +1,11 @@
+import { Config } from '../types';
+
+export const deleteConnection = ({ connectionId, tableName, ddb }: Config) =>
+  ddb
+    .delete({
+      TableName: tableName,
+      Key: {
+        primaryKey: `connectionId:${connectionId}`,
+      },
+    })
+    .promise();
