+++ backend/shared/database/getGroup.ts
@@ -1,0 +1,5 @@
+import { Config, GroupItem } from '../types';
+import { getItem } from './getItem';
+
+export const getGroup = async (groupId: string, config: Config): Promise<GroupItem | void> =>
+  getItem<GroupItem>('groupId', groupId, config.tableName, config.ddb);
